Fix random error when piping a stream
Sometimes, when piping the download of an image to a file, we got the
following error from Resin Request:

	You cannot pipe after data has been emitted from the response.

I was able to capture this in an unit test by adding a delay between the
stream being returned and the client attempting to read data from it.

This means we attempted to pipe the data after it was emitted initially
from the response, thus the error above makes sense.

The solution:

- Quickly pipe the `node-request` stream to a `PassThrough` stream, and
return that instead.

- In order to avoid breaking changes, we had to manually pipe the
`response` event from `node-request` to the `PassThrough` stream.

Side effects:

- I was able to remove the ugly `download.on()` stub to inject the
estimator.

- Cleaned up some tests to use `utils.getStreamData()` instead of
listening to the `data` event and concatenating a stream manually.

- Cleaned up some unit tests that asserted the `Authorization` headers
from the request to consume all the data to trigger a natural `end`
event.