Fix X-Transfer-Length decompression bug
There was a subtle issue about how we treated compressed requests that
contain a `X-Transfer-Length` header.

This header equals the size of the body in it's *compressed* form. We
were currently treating this value as the total value of the download
and consuming the compressed data stream from `node-request`, resulting
in the download being saved without decompression.

The fix is to check whether `Content-Encoding: gzip` is set to determine
if the content is being received in a compressed form, disabling
automatic `gzip` compression on `node-request` and handling
decompression manually with `zlib`.

By handling decompression ourselves, we can choose to perform this step
after or before checking the stream progress, allowing us to decompress
depending on the total size of the progress being the compressed or
decompressed form.