Throw an informative error when the token is expired
If there is a saved token that needs to be refreshed, but it already
expired, the client will get a 'Not Authorized' error.

We can catch this corner case by inspecting the response of the
`GET /whoami` request, which is issued to refresh the token. If the
resulting status code is `401`, we know it represents an scenario where
the token is expired.

Fixes: https://github.com/resin-io/resin-cli/issues/318