Remove saved token if its expired
Currently we throw an error if the token is expired, but keep it there.
The CLI intercepts this error and prints a friendly message asking the
user to login with `resin login`, however this command will attempt to
make some HTTP requests before performing the login, which will throw
the expired error again, and will ask the user to run `resin login` when
actually running `resin login`.